extracted device to a new file
